requirements.txt
claude-agent-sdk>=0.1.58
python-telegram-bot>=21.0
python-dotenv>=1.0.0
pyyaml>=6.0
httpx>=0.27.0
pytest>=8.0
pytest-asyncio>=0.23.0
